optimize: Put css inline
