+++ public/index.html
@@ -8,11 +8,6 @@
         <meta name="description" content="Mithi's Hexapod Robot Simulator" />
         <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
         <title>Mithi's Hexapod Robot Simulator</title>
-        <!--
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="dns-prefetch" href="https://fonts.googleapis.com">
-        -->
-
 
         <link rel="preconnect" crossorigin href="https://www.google-analytics.com">
         <link rel="dns-prefetch" href="https://www.google-analytics.com">
@@ -29,7 +24,11 @@
         <noscript>You need to enable JavaScript to run this app.</noscript>
         <div id="root"></div>
         <style type="text/css">
+        body{font-weight:700;background-color:#222f3e;color:#32ff7e;font-size:3px;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale;margin-left:.3rem;margin-top:5px;touch-action:pan-x pan-y;height:100%}a{color:#ef5777;text-decoration:none;cursor:pointer}a:active,a:hover{color:#ffa801;font-weight:700}h1,h2,h3{margin-top:5px;margin-bottom:5px;margin-left:5px;font-weight:700;font-size:2rem}h2{margin:5px;font-size:1.5rem}h3,p{font-size:.8rem}p{line-height:1.3rem;margin:10px;font-weight:lighter}ul.no-bullet{list-style-type:none;margin:0;padding-left:0}li{margin-top:5px;font-size:.75rem}table{margin-bottom:0;text-align:center;border-collapse:collapse}td,th{border-bottom:1px solid rgba(41,128,185,.5);font-size:12px;line-height:16px;padding:10px}tr:hover{color:#ffa801}.table-container{align-items:center;display:flex;flex-direction:column}.vertical-align{vertical-align:middle}#main{display:flex;margin:0 10px}#sidebar{min-width:350px;flex:1 1;margin-right:15px;box-sizing:border-box}#plot{width:100%}#nav{margin-top:20px}#landing{width:100%}#top-bar{display:grid;grid-template-columns:auto auto auto auto auto auto 1fr;list-style-type:none;margin:0;padding-left:0}.link-icon{font-size:20px;margin:0 .4rem}.text-link{line-height:20px;font-size:13px}.button{color:#ffa801;background-color:#222f3e;border:none;font-size:.6rem;font-weight:700;padding-top:5px;padding-bottom:5px}.button:hover{color:#ef5777;font-weight:700}.text{font-weight:700;font-size:.8rem}.small-icon{font-size:.75rem}.message p{margin:0;font-size:.7rem}.red{color:#e74c3c}.border{margin:5px;padding:5px;border:1px solid #2980b9;border-radius:10px}.label{width:100%;font-size:.6rem;margin-left:5px;opacity:1;white-space:nowrap}.card-header{display:grid;grid-template-columns:auto 1fr}.grid-cols-1{display:grid;grid-template-columns:1fr}.grid-cols-2{display:grid;grid-template-columns:1fr 1fr}.grid-cols-3{display:grid;grid-template-columns:1fr 1fr 1fr}.grid-cols-6{display:grid;grid-template-columns:1fr 1fr 1fr 1fr 1fr 1fr}input{font-size:.6rem;text-align:justify;font-weight:bolder;margin-bottom:10px;width:95%;height:2rem;border-radius:10px;border:1px solid #2980b9;background-color:#2c3e50;color:#2ecc71;box-sizing:border-box;padding-right:7px;text-align:right}input:active,input:hover{color:#ef5777;border:2px solid #ffa801}input::-webkit-inner-spin-button,input::-webkit-outer-spin-button{-webkit-appearance:none;margin:0}input[type=number]{-moz-appearance:textfield}.slider-container{margin:5px}.slider{-webkit-appearance:none;height:4px;border:0;border-radius:5px;background:#ef5777;outline:none;margin:6px}.slider:active,.slider:hover{background:#ffa801;height:8px;margin:4px}.slider::-webkit-slider-thumb{-webkit-appearance:none;appearance:none;width:15px;height:15px;border-radius:50%;background:#2ecc71;cursor:pointer;border:0}.slider::-moz-range-thumb{width:15px;height:15px;border-radius:50%;background:#2ecc71;cursor:pointer;border:0}.switch-container{cursor:pointer;display:flex;align-items:center;margin-left:5px}.switch{position:relative;display:inline-block;width:30px;height:17px;border:1px #2ecc71}.switch+.label{cursor:pointer;width:unset}.switch input{opacity:0;width:0;height:0}.toggle-switch-widget{position:absolute;cursor:pointer;top:0;left:0;right:0;bottom:0;background-color:#ef5777;transition:.4s}.toggle-switch-widget:before{position:absolute;content:"";height:13px;width:13px;left:2px;bottom:2px;background-color:#2c3e50;transition:.4s}input:checked+.toggle-switch-widget{background-color:#2ecc71}input:focus+.toggle-switch-widget{box-shadow:0 0 1px #2ecc71}input:checked+.toggle-switch-widget:before{transform:translateX(13px)}.toggle-switch-widget.round{border-radius:17px}.toggle-switch-widget.round:before{border-radius:50%}@media screen and (min-height:625px) and (min-width:900px){.sidebar{min-height:600px}}@media screen and (max-width:900px){.input{font-size:.7rem}h2{font-size:1.2rem}#sidebar{min-width:325px}#plot{width:100%;height:300px;min-width:250px}td,th{font-size:10px;line-height:15px;padding:10px}}@media screen and (orientation:portrait){.input{font-size:1rem}#main{flex-direction:column-reverse;margin:0}#sidebar{max-width:600px;width:90%;min-width:290px;margin:0 auto}#plot{width:100%;height:400px;box-sizing:border-box;margin:5px 0 0}}@media screen and (max-width:667px) and (orientation:landscape){#top-bar{display:none}#main{flex-direction:column-reverse;margin:0}#sidebar{min-width:225px;margin:0 auto}#plot{margin:8px auto 0;width:90%;height:275px;min-width:250px}table{margin:5px}td,th{font-size:8px;padding:10px 3px}}@media screen and (max-width:767px){input[type=email],input[type=number],input[type=password],input[type=tel],input[type=text]{font-size:16px}}
+        </style>
+        <style type="text/css">
             @import url(https://fonts.googleapis.com/css2?family=Permanent+Marker&display=swap);.button,body,code,input{font-family:Consolas,Monaco,Menlo,Courier,monospace}h1,h2,h3,th{font-family:"Permanent Marker",Consolas,Monaco,Menlo,Courier,monospace}
         </style>
+
     </body>
 </html>
